requirements.txt
flask==3.0.2
playwright==1.42.0
flask-caching==2.1.0
flask-limiter==3.5.0
python-dotenv==1.0.0
gunicorn==21.2.0